backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
boto3
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-jose[cryptography]